+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "allinpay",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "description": "Allinpay Node.js SDK 通联支付 Node.js SDK",
   "main": "index.js",
   "scripts": {
@@ -17,9 +17,11 @@
   },
   "homepage": "https://github.com/calvinxiao/allinpay#readme",
   "dependencies": {
+    "iconv-lite": "^0.4.19",
     "lodash": "^4.17.5",
     "request": "^2.84.0",
-    "request-promise": "^4.2.2"
+    "request-promise": "^4.2.2",
+    "xml2js": "^0.4.19"
   },
   "devDependencies": {
     "mocha": "^5.0.4"
